Compat cleanups, move some compat functions to pyarrow.lib
