+++ python/pyarrow/_fs.pyx
@@ -19,10 +19,9 @@
 
 from cpython.datetime cimport datetime, PyDateTime_DateTime
 
-from pyarrow.compat import frombytes, tobytes
 from pyarrow.includes.common cimport *
 from pyarrow.includes.libarrow cimport PyDateTime_to_TimePoint
-from pyarrow.lib import _detect_compression
+from pyarrow.lib import _detect_compression, frombytes, tobytes
 from pyarrow.lib cimport *
 from pyarrow.util import _stringify_path
