+++ python/pyarrow/io.pxi
@@ -28,8 +28,6 @@ import warnings
 from io import BufferedIOBase, IOBase, TextIOBase, UnsupportedOperation
 
 from pyarrow.util import _is_path_like, _stringify_path
-from pyarrow.compat import (
-    builtin_pickle, frombytes, tobytes, encode_file_path)
 
 
 # 64K
